# monitoring-dashboard/components/org.wso2.ei.dashboard.core/src/main/resources/si/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Siddhi App Service",
    "description" : "This provides information about the REST APIs available for WSO2 Siddhi.",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
    "/siddhi-apps" : {
      "get" : {
        "tags" : [ "artifact" ],
        "summary" : "Lists Siddhi Apps",
        "description" : "Provides the name list of Siddhi Apps that exist.\n",
        "parameters" : [ {
          "name" : "isActive",
          "in" : "query",
          "description" : "Retrieves only active/inactive Siddhi Apps as specified.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The Siddhi Apps are successfully retrieved.",
            "content" : { }
          }
        }
      },
      "put" : {
        "tags" : [ "artifact" ],
        "summary" : "Updates the Siddhi App.",
        "description" : "Updates the Siddhi App.\nRequest \"**siddhiApp** explains the Siddhi Query.\n",
        "requestBody" : {
          "description" : "Siddhi App",
          "content" : {
            "text/plain" : {
              "schema" : {
                "type" : "string"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "The Siddhi App is successfully validated and updated.",
            "content" : { }
          },
          "201" : {
            "description" : "The Siddhi App is successfully validated and saved.",
            "content" : { }
          },
          "400" : {
            "description" : "A validation error occured.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMessageWithCode"
                }
              }
            }
          },
          "500" : {
            "description" : "An unexpected error occured.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMessageWithCode"
                }
              }
            }
          }
        },
        "x-codegen-request-body-name" : "body"
      },
      "post" : {
        "tags" : [ "artifact" ],
        "summary" : "Saves the Siddhi App",
        "description" : "Saves the Siddhi App.\nRequest \"**siddhiApp** explains the Siddhi Query.\n",
        "requestBody" : {
          "description" : "Siddhi App",
          "content" : {
            "text/plain" : {
              "schema" : {
                "type" : "string"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "The Siddhi App is successfully validated and saved.",
            "content" : { }
          },
          "400" : {
            "description" : "A validation error occured.",
            "content" : { }
          },
          "409" : {
            "description" : "A Siddhi App with the given name already exists.",
            "content" : { }
          },
          "500" : {
            "description" : "An unexpected error occured.",
            "content" : { }
          }
        },
        "x-codegen-request-body-name" : "body"
      }
    },
    "/siddhi-apps/{appName}" : {
      "get" : {
        "tags" : [ "artifact" ],
        "summary" : "Retrieves the specified Siddhi App.",
        "description" : "Retrieves the Siddhi App with the specified name.",
        "parameters" : [ {
          "name" : "appName",
          "in" : "path",
          "description" : "The name of the Siddhi App.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The Siddhi App is successfully retrieved.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SiddhiAppContent"
                }
              }
            }
          },
          "404" : {
            "description" : "The Siddhi App specified is not found.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMessageWithCode"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "artifact" ],
        "summary" : "Deletes the Siddhi App.",
        "description" : "Removes the currently deployed Siddhi App with the name specified.\nPath param of **appName** determines name of the Siddhi app.\n",
        "parameters" : [ {
          "name" : "appName",
          "in" : "path",
          "description" : "The name of the Siddhi App.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The Siddhi App is successfully deleted.",
            "content" : { }
          },
          "400" : {
            "description" : "The Siddhi App name provided is invalid.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMessageWithCode"
                }
              }
            }
          },
          "404" : {
            "description" : "The Siddhi App is not found.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMessageWithCode"
                }
              }
            }
          },
          "500" : {
            "description" : "An unexpected error occured.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMessageWithCode"
                }
              }
            }
          }
        }
      }
    },
    "/siddhi-apps/{appName}/status" : {
      "get" : {
        "tags" : [ "artifact" ],
        "summary" : "Fetches the status of the Siddhi App.",
        "description" : "Fetches the status of the Siddhi App.\n",
        "parameters" : [ {
          "name" : "appName",
          "in" : "path",
          "description" : "The name of the Siddhi App.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The state of the Siddhi App is successfully retrieved.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SiddhiAppStatus"
                }
              }
            }
          },
          "404" : {
            "description" : "The Siddhi App specified is not found.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMessageWithCode"
                }
              }
            }
          }
        }
      }
    },
    "/siddhi-apps/{appName}/backup" : {
      "post" : {
        "tags" : [ "state" ],
        "summary" : "Backs up the current state of a Siddhi App.",
        "description" : "Backs up the current state of the specified Siddhi App.\n",
        "parameters" : [ {
          "name" : "appName",
          "in" : "path",
          "description" : "The name of the Siddhi App.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The state of the Siddhi App is successfully persisted.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SiddhiAppRevision"
                }
              }
            }
          },
          "404" : {
            "description" : "The Siddhi App specified is not found.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMessageWithCode"
                }
              }
            }
          },
          "500" : {
            "description" : "An unexpected error occured.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMessageWithCode"
                }
              }
            }
          }
        }
      }
    },
    "/siddhi-apps/{appName}/restore" : {
      "post" : {
        "tags" : [ "state" ],
        "summary" : "Restores the state of a Siddhi App.",
        "description" : "Restores the state of the specified Siddhi App.\n",
        "parameters" : [ {
          "name" : "appName",
          "in" : "path",
          "description" : "The name of the Siddhi App.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "revision",
          "in" : "query",
          "description" : "The revision number of the backup.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "The Siddhi App is successfully restored.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMessage"
                }
              }
            }
          },
          "409" : {
            "description" : "The Siddhi App specified is not found.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMessageWithCode"
                }
              }
            }
          },
          "500" : {
            "description" : "An unexpected error occured.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMessageWithCode"
                }
              }
            }
          }
        }
      }
    },
    "/siddhi-apps/{appName}/revisions" : {
      "delete" : {
        "tags" : [ "State" ],
        "summary" : "Deletes all revisions of the periodic state of a Siddhi Application.",
        "description" : "Deletes all revisions of the periodic state of the specified Siddhi Application.\n",
        "parameters" : [ {
          "name" : "appName",
          "in" : "path",
          "description" : "The name of the Siddhi Application.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "enabledRedeployment",
          "in" : "query",
          "description" : "Whether the redeployment enable or not",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "All revisions of the periodic state of the siddhi application are deleted succussfully.",
            "content" : { }
          },
          "404" : {
            "description" : "The Siddhi Application is not found.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMessageWithCode"
                }
              }
            }
          },
          "500" : {
            "description" : "An unexpected error occured.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMessageWithCode"
                }
              }
            }
          }
        },
        "x-wso2-curl" : "curl -k -X DELETE http://localhost:9090/siddhi-apps/TestSiddhiApp1/revisions",
        "x-wso2-request" : "DELETE http://localhost:9090/siddhi-apps/TestSiddhiApp1/revisions",
        "x-wso2-response" : "HTTP/1.1 200 OK Content-Type: application/json {\"type\":\"success\",\"message\":\"All revisions of periodic state deleted for Siddhi App :TestSiddhiApp1\"}"
      }
    },
    "/siddhi-apps/revisions" : {
      "delete" : {
        "tags" : [ "State" ],
        "summary" : "Deletes all revisions of the periodic state of all Siddhi Applications.",
        "description" : "Deletes all revisions of the periodic state of all Siddhi Applications.\n",
        "parameters" : [ {
          "name" : "enabledRedeployment",
          "in" : "query",
          "description" : "Whether the redeployment enable or not",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "All revisions of the periodic state of all the siddhi applicationa are deleted succussfully.",
            "content" : { }
          },
          "500" : {
            "description" : "An unexpected error occured.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMessageWithCode"
                }
              }
            }
          }
        },
        "x-wso2-curl" : "curl -k -X DELETE http://localhost:9090/siddhi-apps/revisions",
        "x-wso2-request" : "DELETE http://localhost:9090/siddhi-apps/revisions",
        "x-wso2-response" : "HTTP/1.1 200 OK Content-Type: application/json {\"type\":\"success\",\"message\":\"All revisions of periodic state deleted for all Siddhi Apps\"}"
      }
    }
  },
  "components" : {
    "schemas" : {
      "SiddhiAppContent" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "string"
          }
        }
      },
      "SiddhiAppRevision" : {
        "type" : "object",
        "properties" : {
          "revision" : {
            "type" : "string"
          }
        }
      },
      "SiddhiAppStatus" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string"
          }
        }
      },
      "ApiResponseMessage" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
      "ApiResponseMessageWithCode" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "type" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      }
    }
  },
  "x-original-swagger-version" : "2.0"
}